fix: address migration and remaining long-to-int type mismatches
Migration fixes:
- Drop/recreate 5 PostgreSQL views around ALTER COLUMN (PG requires this)
- Remap 224 imported ML training messages with out-of-range bigint IDs
  to sequential negative int values using temp mapping table
- Disable/re-enable FK triggers during remap to avoid constraint violations

Missed type changes found during test run and code review:
- IBotBanHandler/BotBanHandler: long? triggeredByMessageId → int?
- BotModerationService: remove redundant (int) cast on intent.MessageId
- TrustCommand: long? messageId → int?
- WarningEntry (JSONB): long? MessageId → int?

Test fixes:
- BanHandlerTests: 42L/100L literals → int
- CascadeBehaviorTests: FindAsync(5000L) → FindAsync(5000)
- SQL seed files 20/21: remap bigint message IDs to int-range values

+++ TelegramGroupsAdmin.Telegram/Services/Bot/BotModerationService.cs
@@ -554,7 +554,7 @@ public class BotModerationService : IBotModerationService
         {
             var report = new Report(
                 Id: 0, // Assigned by database
-                MessageId: (int)intent.MessageId,
+                MessageId: intent.MessageId,
                 Chat: intent.Chat,
                 ReportCommandMessageId: null,
                 ReportedByUserId: null,
